requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
attrs
av
bitsandbytes
certifi
charset-normalizer
click
cut-cross-entropy
datasets
diffusers
dill
docstring_parser
filelock
frozenlist
fsspec
gitdb
GitPython
hf-xet
hf_transfer
huggingface-hub
idna
importlib_metadata
Jinja2
markdown-it-py
MarkupSafe
mdurl
mpmath
multidict
multiprocess
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging
pandas
peft
pillow
platformdirs
propcache
protobuf
psutil
pyarrow
pydantic
pydantic_core
Pygments
python-dateutil
pytz
PyYAML
qwen-vl-utils
regex
requests
rich
safetensors
sentencepiece
sentry-sdk
setuptools
shtab
six
smmap
sympy
tokenizers
torch
torchvision
tqdm
transformers @ git+https://github.com/huggingface/transformers.git@34133d0a790787739bfc9a42603985de3728ede4
triton
trl @ git+https://github.com/huggingface/trl.git@5787f3bf6365beb0054f984055e4461394933935
typeguard
typing-inspection
typing_extensions
tyro
tzdata
unsloth
unsloth_zoo
urllib3
wandb
wheel
xformers
xxhash
yarl
zipp
